pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-batch"
version = "0.2.0"
description = "Simple batch processing library for Google Gemini API with structured output"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "pymupdf>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gemini_batch*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
markers = [
    "integration: marks tests as integration tests (require live API, deselect with '-m \"not integration\"')",
]
